feat(hostnamegenerator): add zone and namespace variables
