+++ pkg/core/resources/apis/meshexternalservice/hostname/generator.go
@@ -2,14 +2,10 @@ package hostname
 
 import (
 	"context"
-	"fmt"
 	"reflect"
-	"strings"
-	"text/template"
 
 	"github.com/pkg/errors"
 
-	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
 	hostnamegenerator_api "github.com/kumahq/kuma/pkg/core/resources/apis/hostnamegenerator/api/v1alpha1"
 	"github.com/kumahq/kuma/pkg/core/resources/apis/hostnamegenerator/hostname"
 	meshexternalservice_api "github.com/kumahq/kuma/pkg/core/resources/apis/meshexternalservice/api/v1alpha1"
@@ -68,44 +64,5 @@ func (g *MeshExternalServiceHostnameGenerator) GenerateHostname(generator *hostn
 	if !generator.Spec.Selector.MeshExternalService.Matches(es.Meta.GetLabels()) {
 		return "", nil
 	}
-	sb := strings.Builder{}
-	tmpl := template.New("").Funcs(
-		map[string]any{
-			"label": func(key string) (string, error) {
-				val, ok := es.GetMeta().GetLabels()[key]
-				if !ok {
-					return "", errors.Errorf("label %s not found", key)
-				}
-				return val, nil
-			},
-		},
-	)
-	tmpl, err := tmpl.Parse(generator.Spec.Template)
-	if err != nil {
-		return "", fmt.Errorf("failed compiling gotemplate error=%q", err.Error())
-	}
-	type meshedName struct {
-		Name        string
-		DisplayName string
-		Namespace   string
-		Mesh        string
-	}
-	name := es.GetMeta().GetNameExtensions()[model.K8sNameComponent]
-	if name == "" {
-		name = es.GetMeta().GetName()
-	}
-	displayName, ok := es.GetMeta().GetLabels()[mesh_proto.DisplayName]
-	if !ok {
-		displayName = name
-	}
-	err = tmpl.Execute(&sb, meshedName{
-		Name:        name,
-		DisplayName: displayName,
-		Namespace:   es.GetMeta().GetNameExtensions()[model.K8sNamespaceComponent],
-		Mesh:        es.GetMeta().GetMesh(),
-	})
-	if err != nil {
-		return "", fmt.Errorf("pre evaluation of template with parameters failed with error=%q", err.Error())
-	}
-	return sb.String(), nil
+	return hostname.EvaluateTemplate(generator.Spec.Template, es.GetMeta())
 }
